splits the test added #9598 into its own test
This test in ErrorHandlerTest was causing multiple failures in the TCK in servlet and every cloud module
