+++ http-server-tck/src/main/java/io/micronaut/http/server/tck/tests/ErrorHandlerTest.java
@@ -151,21 +151,6 @@ public class ErrorHandlerTest {
     }
 
     @Test
-    void jsonSyntaxErrorBodyAccessible() throws IOException {
-        asserts(SPEC_NAME,
-            CollectionUtils.mapOf(
-                PROPERTY_MICRONAUT_SERVER_CORS_CONFIGURATIONS_WEB_ALLOWED_ORIGINS, Collections.singletonList("http://localhost:8080"),
-                PROPERTY_MICRONAUT_SERVER_CORS_ENABLED, StringUtils.TRUE),
-            HttpRequest.POST("/json/jsonBody", "{\"numberField\": \"textInsteadOf"),
-            (server, request) -> AssertionUtils.assertThrows(server, request,
-                HttpResponseAssertion.builder()
-                    .status(HttpStatus.BAD_REQUEST)
-                    .body("Syntax error: {\"numberField\": \"textInsteadOf")
-                    .build()
-            ));
-    }
-
-    @Test
     void corsHeadersArePresentAfterFailedDeserialisationWhenErrorHandlerIsUsed() throws IOException {
         asserts(SPEC_NAME,
             CollectionUtils.mapOf(
@@ -301,13 +286,6 @@ public class ErrorHandlerTest {
         @Post("/jsonBody")
         String jsonBody(@Valid @Body RequestObject data) {
             return "blah";
-        }
-
-        @Error
-        @Produces(MediaType.APPLICATION_JSON) // it's a lie!
-        @Status(HttpStatus.BAD_REQUEST)
-        String syntaxErrorHandler(@Body @Nullable String body) {
-            return "Syntax error: " + body;
         }
     }
